requirements.txt
torch>=2.2
torchvision>=0.17
numpy>=1.24
matplotlib>=3.8
pillow>=10.0
tqdm>=4.66
pyyaml>=6.0
transformers>=4.38
datasets>=2.18
tokenizers>=0.15
evaluate>=0.4
nltk>=3.8
rouge-score>=0.1
peft>=0.9
accelerate>=0.27
huggingface-hub>=0.21
jupyter>=1.0
ipykernel>=6.0
